backend/requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
asyncio==4.0.0
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0.3
django-cors-headers==4.9.0
django-ninja==1.5.3
djangorestframework==3.16.1
flake8==7.3.0
frozenlist==1.8.0
gunicorn==25.1.0
h11==0.16.0
httptools==0.7.1
idna==3.11
immutabledict==4.3.1
kombu==5.6.2
mccabe==0.7.0
multidict==6.7.1
neo4j==6.1.0
numpy==2.4.2
ortools==9.15.6755
packaging==26.0
pandas==3.0.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.3.0
requests==2.32.5
semantic-version==2.10.0
six==1.17.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
yarl==1.23.0
